% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{wef_articles_list}
\alias{wef_articles_list}
\alias{wef_articles_channels}
\alias{wef_articles_content_types}
\alias{wef_articles_origins}
\alias{wef_collections}
\alias{wef_communities}
\alias{wef_events}
\alias{wef_event_sessions}
\alias{wef_pages}
\alias{wef_projects}
\alias{wef_reports}
\alias{wef_sf_accounts}
\alias{wef_topics}
\alias{wef_streams}
\alias{wef_updates}
\title{Calls}
\usage{
wef_articles_list(pages = 1, n = 25, quiet = !interactive())

wef_articles_channels(channel, pages = 1, n = 25,
  quiet = !interactive())

wef_articles_content_types(type, pages = 1, n = 25,
  quiet = !interactive())

wef_articles_origins(origin, pages = 1, n = 25,
  quiet = !interactive())

wef_collections(pages = 1, n = 25, quiet = !interactive())

wef_communities(pages = 1, n = 25, quiet = !interactive())

wef_events(pages = 1, n = 25, quiet = !interactive())

wef_event_sessions(pages = 1, n = 25, quiet = !interactive())

wef_pages(pages = 1, n = 25, quiet = !interactive())

wef_projects(pages = 1, n = 25, quiet = !interactive())

wef_reports(pages = 1, n = 25, quiet = !interactive())

wef_sf_accounts(pages = 1, n = 25, quiet = !interactive())

wef_topics(pages = 1, n = 25, quiet = !interactive())

wef_streams(pages = 1, n = 25, quiet = !interactive())

wef_updates(pages = 1, n = 25, quiet = !interactive())
}
\arguments{
\item{pages}{Number of pages to retrieve.}

\item{n}{Number of results per page.}

\item{quiet}{If \code{FALSE} prints useful information to the console.}

\item{channel}{Channel to retrieve, see \code{\link{valid_channels}}}

\item{type}{Content type to retrieve, see \code{\link{valid_types}}}

\item{origin}{Origin to retrieve, see \code{\link{valid_origins}}}
}
\value{
A \code{list} the length of the number of pages retrieved.
}
\description{
Get articles
}
\note{
Larger pages (\code{n}) take longer to download and are not necessarily more efficient.
}
\section{Functions}{

\itemize{
  \item{\code{wef_articles_list} Get articles}
  \item{\code{wef_articles_channels} Get articles by channel}
  \item{\code{wef_articles_content_types} Get articles by content-type}
  \item{\code{wef_articles_origins} Get articles by origin}
  \item{\code{wef_collections} Get collections}
  \item{\code{wef_communities} Get communities}
  \item{\code{wef_events} Get events}
  \item{\code{wef_event_sessions} Get events' live sessions}
  \item{\code{wef_pages} Get website pages}
  \item{\code{wef_projects} Get projects}
  \item{\code{wef_reports} Get reports}
  \item{\code{wef_sf_accounts} Get Salesforce accounts}
  \item{\code{wef_topics} Get topics}
  \item{\code{wef_streams} Get streams}
  \item{\code{wef_updates} Get updates}
}
}

\examples{
articles <- wef_articles_list()

}
