% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_channels}
\alias{valid_channels}
\alias{valid_types}
\alias{valid_origins}
\title{Valid}
\usage{
valid_channels()

valid_types()

valid_origins()
}
\value{
a \code{list}
}
\description{
Get valid arguments for other calls.
}
\examples{
valid_channels()

}
