% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{wef_articles}
\alias{wef_articles}
\title{Articles}
\usage{
wef_articles(pages = 1, n = 25, quiet = !interactive())
}
\arguments{
\item{pages}{Number of pages to retrieve.}

\item{n}{Number of results per page.}

\item{quiet}{If \code{FALSE} prints useful information to the console.}
}
\value{
A \code{list} of articles the length of the number of pages retrieved.
}
\description{
Get articles
}
\note{
Larger pages (\code{n}) take longer to download and are not necessarily more efficient.
}
\examples{
articles <- wef_articles()

}
