% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callAPI.R
\name{wef_call}
\alias{wef_call}
\title{Call API}
\usage{
wef_call(url, pages = 1, n = 25, quiet = !interactive(),
  sleep = 0.5)
}
\arguments{
\item{url}{URL to call.}

\item{pages}{Number of pages to retrieve.}

\item{n}{Number of results per page.}

\item{quiet}{If \code{FALSE} prints useful information to the console.}

\item{sleep}{Pause in seconds to mark between each API call (\code{pages}).}
}
\value{
A \code{list}.
}
\description{
Call API endpoint.
}
\examples{
acc <- wef_sf_accounts()

rand_acc <- sample(acc[[1]]$attributes$api_endpoint, 1)

data <- wef_call(rand_acc)

}
